Refactor shared uber jar generation code into common subclass.
